implment dhcp discover

+++ kernel/net/addr.c
@@ -5,8 +5,8 @@
 
 
 void ip2str(string_t str,ipv4_addr_t * ip){
-    sprintf(str,"%d.%d.%d.%d",ip->n[0],ip->n[1],ip->n[2],ip->n[3]);
+    sprintf(str,"%d.%d.%d.%d",(uint32_t)ip->n[0],(uint32_t)ip->n[1],(uint32_t)ip->n[2],(uint32_t)ip->n[3]);
 }
 void mac2str(string_t str,eth_addr_t * mac){
-    sprintf(str,"%x:%x:%x:%x:%x:%x",mac->n[0],mac->n[1],mac->n[2],mac->n[3],mac->n[4],mac->n[5]);
+    sprintf(str,"%x:%x:%x:%x:%x:%x",(uint32_t)mac->n[0],(uint32_t)mac->n[1],(uint32_t)mac->n[2],(uint32_t)mac->n[3],(uint32_t)mac->n[4],(uint32_t)mac->n[5]);
 }
